# app/web/api/openapi.generated.json
{
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "in": "cookie",
        "name": "session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "OpenAPI description for the Stellar backend.",
    "title": "Stellar API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
    "/api/auth": {
      "get": {
        "responses": {
          "200": {
            "description": "Authenticated user"
          },
          "401": {
            "description": "Not authorized"
          }
        },
        "summary": "Get currently authenticated user",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/login": {
      "get": {
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "summary": "GET /api/auth/login"
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "password": {
                    "type": "string"
                  },
                  "username": {
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Authenticated user"
          },
          "401": {
            "description": "Invalid credentials"
          }
        },
        "summary": "Authenticate user",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/logout": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Logout response"
          }
        },
        "summary": "Logout current user",
        "tags": [
          "Authentication"
        ]
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Logout response"
          }
        },
        "summary": "Logout current user",
        "tags": [
          "Authentication"
        ]
      }
    },
...
    "/api/galaxy/batch_initialize": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "galaxy_name": {
                    "type": "string"
                  },
                  "systems": {
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "description": "Initialized systems"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Not authorized"
          }
        },
        "summary": "Batch initialize galaxy systems",
        "tags": [
          "Galaxy"
        ]
      }
    },
    "/api/galaxy/create": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "properties": {
                    "type": "string"
                  },
                  "sectors": {
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "description": "Created galaxy"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Not authorized"
          },
          "409": {
            "description": "Conflict"
          }
        },
        "summary": "Initialize a galaxy",
        "tags": [
          "Galaxy"
        ]
      }
    },
...
    "/api/galaxy/{name}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Galaxy details"
          },
          "404": {
            "description": "Galaxy not found"
          }
        },
        "summary": "Get galaxy details",
        "tags": [
          "Galaxy"
        ]
      }
    },
    "/api/galaxy/{name}/systems": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "mine",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "galaxy_name": {},
                    "properties": {},
                    "systems": {}
                  },
                  "required": [
                    "galaxy_name",
                    "systems",
                    "properties"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Systems in galaxy"
          }
        },
        "summary": "List systems for a galaxy",
        "tags": [
          "Galaxy"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:9000"
    }
  ]
}